add. column

+++ src/main/java/api/lemonico/resource/WarehouseStoreResource.java
@@ -45,8 +45,8 @@ public class WarehouseStoreResource
     /** 更新日時 */
     private final LocalDateTime modifiedAt;
 
-    /** 削除フラグ（退会から一定時間経過後に削除状態になる） */
-    private final Integer isDeleted;
+    /** 削除フラグ（0: 未削除 1: 削除済） */
+    private final Byte isDeleted;
 
     /**
      * 指定したエンティティを使用して、リソースを構築します。
